requirements.txt
langchain==0.3.7
langchain-google-genai==2.0.5
langchain-community==0.3.7
faiss-cpu==1.9.0
pypdf==5.1.0
python-dotenv==1.0.1
streamlit==1.40.2
